fix device bug

+++ src/semi_supervised/semi_supervised_trainer.py
@@ -197,7 +197,7 @@ class SemiSupervisedTrainer(GenerativeTrainer):
                 progress_bar.update(data.size(0))
 
                 correct += (torch.argmax(net_args[2][-1], dim=1) == labels.to(device)).sum().float()
-                total += len(labels).float()
+                total += len(labels)
 
         print(f"===============> Epoch {epoch}; Accuracy: {correct/total}")
         # print(net.means)
